+++ timer.h
@@ -1,7 +1,7 @@
 #ifndef TIMER_H
 #define TIMER_H
 
-volatile extern uint16_t timerCounter;
+volatile extern uint32_t timerCounter;
 
 void init_timer();
 void reset_timer();
